Cleaning Sequences
There are multiple ways of sequencing the steps in your data cleaning process. We've used one so far, once in Lesson 1 and once in the Data Cleaning Process example video in this lesson. The Define, Code, and Test Markdown headers were used once in this sequence, with multiple definitions, cleaning operations, and tests under each header, respectively.

It looked like this: